Update vite-issues.mdx yalan2ny
For better use of wallets, we add Web3 auth to the wallet browser extension so that wallet connect scanning can be done better and easier.  My problems: wallet connect is not working, metamsk does not show asset balance and none of the wallets are able to send or receive, I am trying to solve the problems by requesting Web3 auth integration.