Update how-web3auth-works.mdx yal2ny
We will add Web3 auth, which is complementary, to the wallet browser extension so that wallet connect scanning can be done better and more easily.  My problems: wallet connect is not working, metamsk does not show asset balance and none of the wallets are able to send or receive, I am trying to solve the problems by requesting Web3 auth integration.